Give parent tracking more context
Linker need to propagate links from arguments of function pointers to the actual
functions. Similarly for return values of function pointers. For that it needs
to know, whether any given symbol is a return value or argument, and if the
latter, which. To that end, add that tracking to call graphs as well as the
callgraph parser.