calgary: Mark function pointer nodes
Linker needs to know which nodes are pointers to functions, so that it can
forward callees of their arguments to the actual function that the pointer might
reference. These nodes also aren't declarations, so don't mark them so.

Improve tracking of context so that these nodes are properly marked as static
only if they really aren't supposed to be visible outside of the current
callgraph.