calgary: Support statement expressions
A statement expression such as:

    void bar(void);
    static void foo(void) {
        void (*ptr)(void) = ({ ...; ...; &bar; });
    }

Is compiled into:

    modify_expr(ptr, target_expr(tmp,, ({...; ...; modify_expr(tmp, &bar)})))

The variable "tmp" is an ad-hoc temporary variable, which oesn't have a function
context. Therefore edges are not propagated from this variable when call graph
is merged, and the variable remains in the call graph. It is also nameless,
which results in weird nodes such as:

    1792 (1) @decl bar
    1793 (2) @static foo
    1795 (3) @static @var foo():: 1792

Recognize that nameless variables are local as well and should be thereore
propagated.