calgary: Support call of a result of a call
Instead of dealing with it right away, defer handling call-of-a-call case
to the code that actually sees the callee node. This maps better to the
tree-walk model. An alternative would be returning a set of ultimate
callees with their depths to from walk_call_expr(), but that would be
awkward.

This doesn't include the linker support yet.