calgary: Inline get_callee into walk_call_expr
One branch of a ternary operator may be wrapped in a NOP_EXPR. NOP_EXPR is not
handled by walk_call_expr, but by get_callee. Correspondingly, if that branch
contains a call, that call is not recorded.

NOP_EXPR therefore needs to be processed in walk_call_expr. In order for the
typecast calls to work as well, the type deduced from the NOP_EXPR needs to be
passed to the recursive invocations of walk_call_expr. This obviates the need
for a separate class callee, so drop it.

Eventually, it seems like there are more and more opcodes that need special
handling in walk_call_expr, and it's confusing to have another recursive walker
under get_callee. Just inline all the other nodes to walk_call_expr as well.

To make the separation between the walking and the handling clear, put the
latter into a separate function, handle_call, called from walk_call_expr.

Add to call_cond_expr.c/.cg stuff that would crash calgary before this commit.